+++ sdk/core/iOS/MoEReactBridge/MoEngageReactConstants.m
@@ -44,3 +44,4 @@ NSString* const kPushTokenGenerated = @"MoEPushTokenGenerated";
 NSString* const kPushClicked        = @"MoEPushClicked";
 NSString* const kPermissionResult   = @"MoEPermissionResult";
 NSString* const kLogoutComplete     = @"MoELogoutComplete";
+NSString* const kAuthenticationError = @"MoEAuthenticationError";